본인 글만 수정, 삭제 가능

+++ src/main/java/com/jnh/springboottest/article/ArticleController.java
@@ -1,10 +1,13 @@
 package com.jnh.springboottest.article;
 
+import com.jnh.springboottest.member.Member;
+import com.jnh.springboottest.member.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.util.List;
 
 @Controller
@@ -13,6 +16,7 @@ import java.util.List;
 public class ArticleController {
 
     private final ArticleService articleService;
+    private final MemberService memberService;
 
     @GetMapping("/list")
     public String showList(Model model){
@@ -28,8 +32,14 @@ public class ArticleController {
     }
 
     @PostMapping("/create")
-    public String create(@RequestParam String title, @RequestParam String content){
-        articleService.write(title, content);
+    public String create(@RequestParam String title, @RequestParam String content, Principal principal) {
+        Member member = null;
+        try{
+            member = memberService.findByUsername(principal.getName());
+        }catch (NullPointerException nullPointerException){
+            return "article_create";
+        }
+        articleService.write(title, content, member);
         return "redirect:/article/list";
     }
 
@@ -45,31 +55,43 @@ public class ArticleController {
     }
 
     @GetMapping("/modify/{id}")
-    public String showModify(@PathVariable int id, Model model){
-        try{
+    public String showModify(@PathVariable int id, Model model, Principal principal) {
+        try {
             Article article = articleService.readOne(id);
-            model.addAttribute(article);
-        }catch (Exception e){
+            if (!article.getMember().getUsername().equals(principal.getName())) {
+                return "redirect:/article/detail/" + id;
+            }
+            model.addAttribute("article", article);
+        } catch (Exception e) {
             e.printStackTrace();
         }
         return "article_create";
     }
 
     @PostMapping("/modify/{id}")
-    public String modify(@PathVariable int id, @RequestParam String title, @RequestParam String content){
-        try{
+    public String modify(@PathVariable int id, @RequestParam String title,
+                         @RequestParam String content, Principal principal) {
+        try {
+            Article article = articleService.readOne(id);
+            if (!article.getMember().getUsername().equals(principal.getName())) {
+                return "redirect:/article/detail/" + id;
+            }
             articleService.modify(id, title, content);
-        }catch (Exception e){
+        } catch (Exception e) {
             e.printStackTrace();
         }
         return "redirect:/article/detail/" + id;
     }
 
     @PostMapping("/remove/{id}")
-    public String remove(@PathVariable int id){
-        try{
+    public String remove(@PathVariable int id, Principal principal) {
+        try {
+            Article article = articleService.readOne(id);
+            if (!article.getMember().getUsername().equals(principal.getName())) {
+                return "redirect:/article/detail/" + id;
+            }
             articleService.remove(id);
-        }catch (Exception e){
+        } catch (Exception e) {
             e.printStackTrace();
         }
         return "redirect:/article/list";
